Move methods for getting prefab and block names to the shop ui screen,
